crud operations completed

+++ src/main/java/com/codejam/demo/repository/PersonalInformationRepository.java
@@ -3,5 +3,5 @@ package com.codejam.demo.repository;
 import com.codejam.demo.model.PersonalInformationEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonalInformationRepository extends CrudRepository<PersonalInformationEntity, Long> {
+public interface PersonalInformationRepository extends CrudRepository<PersonalInformationEntity, Integer> {
 }
