+++ src/main/java/com/codejam/demo/repository/ScheduleRepository.java
@@ -3,5 +3,7 @@ package com.codejam.demo.repository;
 import com.codejam.demo.model.ScheduleEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ScheduleRepository extends CrudRepository<ScheduleEntity, Long> {
+import javax.persistence.criteria.CriteriaBuilder;
+
+public interface ScheduleRepository extends CrudRepository<ScheduleEntity, Integer> {
 }
